requirements.txt
fastapi
uvicorn[standard]
python-dotenv
lancedb
pyarrow
pypdf
openai
